requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.23.0
scipy>=1.10.0

# Geospatial Libraries
geopandas>=0.12.0
shapely>=2.0.0
osmnx>=1.9.0
pyproj>=3.5.0
rtree>=1.0.1

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
imbalanced-learn>=0.10.1

# Model Interpretation
shap>=0.42.0
lime>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.6

# Development Tools
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.1.0
pylint>=2.16.0
flake8>=6.0.0
mypy>=1.1.0
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.28.2
pint>=0.20.0

# kaggle dataset
kagglehub>=0.3.13

# Database (if needed)
sqlalchemy>=2.0.0
